[*] tk: restructure matmul per-GPU + rehome gfx9 asm intrinsics
Split the monolithic matmul.rs into a per-GPU tree and lift the inline-asm
scheduling machinery off the Kernel into explicit primitives, so gfx942 (CDNA3)
and gfx1151 (RDNA3.5) can evolve independently.

- matmul.rs -> matmul/{mod,common}.rs + amd/{mod,gfx942,gfx1151}.rs. gfx942 holds
  the asm HK-pipeline microkernel (gemm_core_hk) + the 256²/128² configs; common
  holds the arch-neutral gemm_core / split-K; gfx1151 the RDNA config.
- asm.rs -> arch/gfx9.rs, declared pub(crate) (off the public API): the CDNA/wave64
  MFMA-pipeline intrinsics (s_setprio, s_waitcnt_lgkmcnt, s_barrier_bare,
  sched_barrier, wave_phase_barrier) now live beside the arch caps they belong with.
- Kernel: drop the asm_mfma/asm_gather/asm_prefetch mode flags; each asm op is an
  explicit per-call primitive (mma_abt_asm, gather_local_asm, stage/commit_*_asm).
- gemm_core_hk drain cleanup: match HK's drain topology (7 -> 4 lgkmcnt(0)/iter) by
  making the two post-MMA boundaries bare and gathering C5's operand two clusters
  ahead. Verified neutral vs the asm-pin (disasm strict-subset; isolated 808 TF
  @8192; route A/B identical) — the schedule already was HK's.
- codegen: carry fast-math flags (nsz arcp contract afn) on the WMMA/MFMA intrinsic
  to match the rest of the fp ops (+ declare auto-scan parser strips leading FM
  flags before the ret-ty).
- Remove the golden test module; add the 128² HK variant + tests. Drop dead
  experimental code (iglp_opt, the s_setprio_after fan-in).

Build/clippy/fmt clean; 6/6 matmul + 7/7 FA device numerics pass.
